ML advanced, assignment 1
In the first task we will look at a supervised scenario where we work with a model of a specific relationship
between two different domains. The second task we look at how we can perform unsupervised learning and learn a new representation of the data. Finally we will end with a look at how we can approach model selection.